Reset full state when user leaves the room
